scripts/requirements.txt
# Para count_questions_three_languages.py
firebase-admin>=6.0.0
python-dotenv>=1.0.0
